Case 8332: Fix int type for chunk_id, where loss of precision was forcing re-reading index chunk data for every record of chr/start/stop index mapped fields. (slow.....)

+++ src/tsf.h
@@ -252,7 +252,7 @@ typedef struct tsf_chunk {
   tsf_chunk_header header;
 
   int record_count;
-  int chunk_id;
+  int64_t chunk_id;
   tsf_value_type value_type;
   char* chunk_data; // Just a bunch of bytes
   int chunk_bytes; //length of chunk_data
